Create omxplayer-sync-tests.py
This will be the main python script that gets called to attempt to synchronize the video feeds across multiple raspis. It is extremely important that one raspi be devoted exclusively to the launching of the threads, because I shit you not omxplayer uses up 99.9% of the raspi's resources, and if you so much as issue an 'echo hello world' statement in the background, the fps will drop from 31 to 2...but I digress. 

Alternatively, you can launch these threads from a desktop. As all that you need is Python, your standard linux desktop should not fuck up your life too much. That being said, linux has on more than one occasion reinforced this author's belief in the principle of Murphy's Law.

In either case, the tests must be run with one raspi/desktop launching the threads (here called a 'raspi server') and 1 or more raspis launching omxplayer (here called a 'raspi client'). Although it is possible to run this test using only one raspi client, it defeats the purpose of testing across multiple clients.