Update omxplayer-sync-tests.sh
Finalized script with commenting to make it easy for all users to modify. This program essentially waits, with frequency 0.1s until the desired launch time is encountered. Not only did this prove to be an inaccurate way to carry this out, but it turns out that even if two raspis call omxplayer at the same time, there is still no guarantee that the movies will launch at the same time.